fix/context menu implementation
