+++ src/components/solves/ModalSolve.tsx
@@ -2,7 +2,6 @@ import updateSolve from "@/lib/updateSolve";
 import { useSolvesStore } from "@/store/SolvesStore";
 import { useTimerStore } from "@/store/timerStore";
 import formatTime from "@/lib/formatTime";
-import moveSolve from "@/lib/moveSolve";
 import { ScrambleDisplay } from "@/components/scramble-display/index";
 import useEscape from "@/hooks/useEscape";
 import { format } from "date-fns";
@@ -13,20 +12,18 @@ import { Solve } from "@/interfaces/Solve";
 import { SolveTab } from "@/interfaces/types/SolveTabs";
 import { useTranslations } from "next-intl";
 import {
-  ArchiveBoxArrowDownIcon,
   CalendarDaysIcon,
   ChevronDownIcon,
   ChevronUpIcon,
   CubeTransparentIcon,
-  DocumentDuplicateIcon,
   EllipsisHorizontalIcon,
   StarIcon,
-  TrashIcon,
 } from "@heroicons/react/24/solid";
 import {
   StarIcon as StarIconO,
   ChatBubbleBottomCenterIcon as ChatBubbleBottomCenterIconO,
 } from "@heroicons/react/24/outline";
+import ContextMenu from "./ContextMenu";
 
 export default function ModalSolve({ currentTab }: { currentTab: SolveTab }) {
   const [showOptions, setShowOptions] = useState<boolean>(false);
@@ -44,28 +41,6 @@ export default function ModalSolve({ currentTab }: { currentTab: SolveTab }) {
   useClickOutside(submenuRef, () => setShowOptions(false));
 
   useEscape(() => setStatus(false));
-
-  const handleMove = async (solve: Solve, currentTab: SolveTab) => {
-    if (!selectedCube) return;
-    const updatedCube = await moveSolve({
-      solve,
-      selectedCube,
-      type: currentTab,
-    });
-    mergeUpdateSelectedCube(updatedCube, cubes);
-    setStatus(false);
-  };
-
-  const handleDelete = async (solve: Solve) => {
-    if (!selectedCube) return;
-    const updatedCube = await updateSolve({
-      selectedCube: selectedCube,
-      solveId: solve.id,
-      type: "DELETE",
-    });
-    mergeUpdateSelectedCube(updatedCube, cubes);
-    setStatus(false);
-  };
 
   const handlePlusTwo = async (solve: Solve) => {
     if (!selectedCube) return;
@@ -98,13 +73,6 @@ export default function ModalSolve({ currentTab }: { currentTab: SolveTab }) {
       type: "BOOKMARK",
     });
     mergeUpdateSelectedCube(updatedCube, cubes);
-    setStatus(false);
-  };
-
-  const handleCopyToClipboard = async (text: string) => {
-    if ("clipboard" in navigator) {
-      await navigator.clipboard.writeText(text);
-    }
     setStatus(false);
   };
 
@@ -230,57 +198,13 @@ export default function ModalSolve({ currentTab }: { currentTab: SolveTab }) {
                 </div>
               </div>
               {/* options menu */}
-              <AnimatePresence>
-                {showOptions && (
-                  <motion.div
-                    initial={{ y: 0, scale: 0.9, opacity: 0 }}
-                    animate={{ y: 0, scale: 1, opacity: 1 }}
-                    exit={{ x: 0, scale: 0.9, opacity: 0 }}
-                    ref={submenuRef}
-                    className="absolute flex flex-col w-32 gap-3 py-2 mt-1 bg-white rounded-md"
-                  >
-                    <div
-                      className="flex items-center gap-1 py-1 transition duration-200 ps-2 hover:text-neutral-500 hover:cursor-pointer"
-                      onClick={() =>
-                        currentTab === "Session"
-                          ? handleMove(solve, "Session")
-                          : handleMove(solve, "All")
-                      }
-                    >
-                      <div className="w-4 h-4">
-                        <ArchiveBoxArrowDownIcon className="w-4 h-4 fill-none stroke-black" />
-                      </div>
-                      <div>
-                        {currentTab === "Session"
-                          ? t("archive")
-                          : t("unarchive")}
-                      </div>
-                    </div>
-                    <div
-                      className="flex items-center gap-1 py-1 transition duration-200 ps-2 hover:text-neutral-500 hover:cursor-pointer"
-                      onClick={() =>
-                        handleCopyToClipboard(
-                          `[${formatTime(solve.time)}s] - ${solve.scramble}`
-                        )
-                      }
-                    >
-                      <div className="w-4 h-4">
-                        <DocumentDuplicateIcon className="w-4 h-4 fill-none stroke-black" />
-                      </div>
-                      <div>{t("copy")}</div>
-                    </div>
-                    <div
-                      className="flex items-center gap-1 py-1 transition duration-200 ps-2 hover:text-neutral-500 hover:cursor-pointer"
-                      onClick={() => handleDelete(solve)}
-                    >
-                      <div className="w-4 h-4">
-                        <TrashIcon className="w-4 h-4 fill-none stroke-black" />
-                      </div>
-                      <div>{t("remove")}</div>
-                    </div>
-                  </motion.div>
-                )}
-              </AnimatePresence>
+              {showOptions && (
+                <ContextMenu
+                  submenuRef={submenuRef}
+                  currentTab={currentTab}
+                  solve={solve}
+                />
+              )}
             </motion.div>
           </div>
         ) : null}
